// cat-launcher/src/generated-types/GameReleaseStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * The possible installation statuses of a game release.
 */
export type GameReleaseStatus = "NotAvailable" | "NotDownloaded" | "Corrupted" | "NotInstalled" | "ReadyToPlay" | "Unknown";